+++ packages/metatable/src/__tests__/utils.spec.ts
@@ -1,6 +1,6 @@
  import { filterColumnPaths, unsetAllSortFormValues, toMetaFilters } from '../utils';
 import {Columns, IMetaFiltersArgs, MetaField} from '@falsy/metacore';
- import {getDateRangeFilter, getTextFilter, getTernaryFilter, columnBuilder} from "metahelpers";
+ import {getDateRangeFilter, getTextFilter, getTernaryFilter, columnBuilder} from "metaforms";
 
 const customerIdFilterForm: MetaField[] = [
   {
@@ -138,7 +138,7 @@ const mockedColumns1: Columns<'number' | 'timestamp' | 'string' | 'boolean'> = {
       {
         name: 'submit',
         type: 'submit',
-        label: 'Uložit',
+        label: 'Save',
       },
     ],
   },