+++ packages/metaforms/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-alpha.14](https://github.com/flsy/metaforms/compare/metaforms@2.0.0-alpha.13...metaforms@2.0.0-alpha.14) (2022-09-03)
+
+**Note:** Version bump only for package metaforms
+
+
+
+
+
+# [2.0.0-alpha.13](https://github.com/flsy/metaforms/compare/metaforms@2.0.0-alpha.12...metaforms@2.0.0-alpha.13) (2022-09-03)
+
+**Note:** Version bump only for package metaforms
+
+
+
+
+
+# [2.0.0-alpha.12](https://github.com/flsy/metaforms/compare/metaforms@2.0.0-alpha.11...metaforms@2.0.0-alpha.12) (2022-09-02)
+
+**Note:** Version bump only for package metaforms
+
+
+
+
+
 # 2.0.0-alpha.11 (2022-08-18)
 
 **Note:** Version bump only for package metaforms